support_bundle_20250922_215921/files/requirements.txt
requests
beautifulsoup4
aiohttp
pdfminer.six
pymupdf
pytesseract
pillow
openai
sentence-transformers
faiss-cpu
chromadb

fastapi==0.114.2
uvicorn[standard]==0.30.6
faiss-cpu==1.8.0.post1
numpy==1.26.4
openai==1.108.1
